requirements.txt
pillow
webcolors